Working refactored WebGPU prototype
