+++ src/topsy/shaders/sph.wgsl
@@ -1,6 +1,8 @@
 struct TransformParams {
     transform: mat4x4<f32>,
     scale_factor: f32,
+    clipspace_size_min: f32,
+    clipspace_size_max: f32
 };
 
 @group(0) @binding(0)
@@ -9,7 +11,7 @@ var<uniform> trans_params: TransformParams;
 
 struct VertexInput {
    @location(0) pos: vec4<f32>, // NB w is used for the smoothing length
-   @location(1) offset: vec2<f32>, // the offset is used to expand into a quad
+   @builtin(vertex_index) vertexIndex: u32
 }
 
 struct VertexOutput {
@@ -23,19 +25,45 @@ struct FragmentOutput {
     @location(0) color: vec4<f32>
 }
 
+
 @vertex
 fn vertex_main(input: VertexInput) -> VertexOutput {
+    // triangle position offsets for making a square of 2 units side length
+    var posOffset = array<vec2<f32>, 6>(
+                vec2(-1.0, -1.0),
+                vec2(-1.0, 1.0),
+                vec2(1.0, 1.0),
+                vec2(1.0, -1.0),
+                vec2(-1.0, -1.0),
+                vec2(1.0, 1.0)
+              );
+
+    // corresponding texture coordinates
+    var texCoords = array<vec2<f32>, 6>(
+                vec2(0.0, 0.0),
+                vec2(0.0, 1.0),
+                vec2(1.0, 1.0),
+                vec2(1.0, 0.0),
+                vec2(0.0, 0.0),
+                vec2(1.0, 1.0)
+              );
+
     var output: VertexOutput;
+
+    // smoothing length is stored in w
     var size = trans_params.scale_factor*input.pos.w;
-    // w is hijacked for the smoothing length. However since it will be used in the
-    // transformation and clipping, we now need to set it to the 'standard' 1.0.
-    output.pos = input.pos;
-    output.pos.w = 1.0;
-    output.pos = (trans_params.transform * output.pos);
-    output.z = output.pos.z;
-    output.pos.x+=size*(input.offset.x - 0.5);
-    output.pos.y+=size*(input.offset.y - 0.5);
-    output.texcoord = input.offset.xy;
+
+    if(size<trans_params.clipspace_size_min || size>trans_params.clipspace_size_max) {
+        output.pos.z = 100.0; // discard the vertex
+    } else {
+        // perform transformation
+        output.pos = input.pos;
+        output.pos.w = 1.0;
+        output.pos = (trans_params.transform * output.pos);
+        output.z = output.pos.z;
+        output.pos += vec4<f32>(size*posOffset[input.vertexIndex],0.0,0.0);
+        output.texcoord = texCoords[input.vertexIndex];
+    }
     return output;
 }
