+++ src/topsy/util.py
@@ -1,0 +1,4 @@
+def load_shader(name):
+    from importlib import resources
+    with open(resources.files("topsy.shaders") / name, "r") as f:
+        return f.read()
